Check consistency between persisted snapshots
A previous PR added an hourly cronjob to check consistency. This didn't
check consistency as expected. After each invocation, a log file is
written to disk with the previous checkpoint. Consistency is checked
between the previous and current checkpoints. Without persistence, the
job would check consistency between the same checkpoints.

To implement persistence, we use a GitHub Action that downloads
previously uploaded artifacts. Persistence between workflows isn't
natively supported.

This also changes the persisted format to be a signed checkpoint rather
than just the size and root hash, which means we can verify it.

This also switches the default URL to be rekor.sigstore.dev, as api was
deprecated.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>