Immediately tick to avoid waiting an interval
Fixes https://github.com/sigstore/rekor-monitor/issues/588

With a recent refactor where we added support for catching a signal,
this caused the ticker to wait a full interval before running the
consistency check, which meant that runtime went from immediate to 5 min
by default. This fixes the bug by moving the for-select logic to the end
of the loop, a solution recommended online.

Signed-off-by: Hayden Blauzvern <hblauzvern@google.com>