Split the consistency check and the checkpoint writing
Before this patch, the consistency check function was also writing the
new checkpoint to file, before searching for matching identities. If
something would happen during the identity matching, by the time you
re-run the monitor you'd start from the new checkpoint, rather than the
previous one, skipping the entries that were part of the last checkpoint
"diff".

This patch splits the checkpoint consistency check and the storing of
the new checkpoint, so that it is written only after all identities have
been searched. If an error happens during the identity search, the new
checkpoint would not be written to file and the next monitor run would
start again from the old checkpoint (at the expense of extra time
required to re-analyze all the entries).

Signed-off-by: Riccardo Schirone <riccardo.schirone@trailofbits.com>